requirements.txt
# Core Scientific Computing
numpy>=1.23  # Allow numpy 2.x if available (install with --only-binary to avoid source builds)
pandas>=1.5
scipy>=1.10

# Causal Inference & Graph Theory
rustworkx>=0.12.0
networkx>=3.0
dowhy>=0.8
causal-learn>=0.1.3.9
tigramite>=5.2.0

# Optimization
# Note: cvxpy requires C++ build tools on Windows (Microsoft Visual C++ 14.0+)
# Install separately if needed: pip install cvxpy
# Or install build tools: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# cvxpy>=1.4  # Optional - code handles missing cvxpy gracefully
optuna>=3.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
numba>=0.58

# Data Sources & APIs
yfinance>=0.2.0
pycoingecko>=3.0
pytrends>=4.8.0
praw>=7.0
shodan>=1.30.0
geoip2>=4.7.0

# Trading & Exchanges
ccxt>=4.0
websockets>=11.0

# Blockchain
web3>=6.0

# Visualization
matplotlib>=3.5
seaborn>=0.11
pillow>=10.0
reportlab>=4.0.0

# Utilities
requests>=2.28
loguru>=0.7
python-dotenv>=1.0
python-dateutil>=2.8
pyyaml>=6.0
tqdm>=4.65.0

# UI
rich>=13.0

# Web Framework
flask>=2.3.0

# Caching
redis>=5.0

# Data Validation & Serialization
pydantic>=2.0
toml>=0.10.2

# MCP (Model Context Protocol)
mcp>=0.9.0
openai>=1.0.0

# Framework
swarms
litellm>=0.1

# System Monitoring
psutil>=5.9.0

# LRM Training (optional)
transformers>=4.39.0
datasets>=2.16.0
accelerate>=0.26.0
peft>=0.9.0

# Optional Dependencies (uncomment if needed)
# cvxpy>=1.4  # Requires C++ build tools on Windows

# Image Annotation System
scikit-image>=0.21.0
pywavelets>=1.4.0
opencv-python>=4.8.0  # Install with --only-binary to use pre-built wheels
sympy>=1.12
z3-solver>=4.13.0
filterpy>=1.4.5